+++ HashFuncs.h
@@ -1,6 +1,6 @@
 #ifndef HASH_FUNCS_H
 #define HASH_FUNCS_H
 
-unsigned int MurmurHash(const void* hashingArr, const size_t length, const unsigned int seed);
+uint64_t MurmurHash(const void* hashingArr, const size_t length, const uint64_t seed);
 
 #endif // HASH_FUNCS_H
